+++ scatter/report/components/charts/ChartCore.tsx
@@ -1,5 +1,7 @@
 'use client'
 import dynamic from 'next/dynamic'
+import React from 'react'
+import {LoadingBar} from '@/components/LoadingBar'
 
 export const ChartCore = dynamic(async () => {
   const Plotly = await import('plotly.js/lib/core')
@@ -9,4 +11,4 @@ export const ChartCore = dynamic(async () => {
   const createPlotlyComponent = (await import('react-plotly.js/factory')).default
   Plotly.register([Scatter, Sunburst, Treemap])
   return createPlotlyComponent(Plotly)
-}, { ssr: false })
+}, { ssr: false, loading: () => <LoadingBar loaded={100} max={100} isVisualizing /> })
